adapt solution for Gobra interfaces
