+++ src/main/scala/viper/silver/ast/Ast.scala
@@ -6,15 +6,10 @@
 
 package viper.silver.ast
 
-import viper.silver.ast
-import viper.silver.ast.EdgeType.EdgeType
-import viper.silver.ast.JoinType.JoinType
 import viper.silver.ast.pretty.FastPrettyPrinter
 import viper.silver.ast.utility._
 import viper.silver.ast.utility.rewriter.Traverse.Traverse
 import viper.silver.ast.utility.rewriter.{Rewritable, StrategyBuilder, Traverse}
-import viper.silver.dependencyAnalysis.DependencyType.{ExplicitAssertion, ExplicitAssumption, MethodCall, Rewrite, SourceCode}
-import viper.silver.dependencyAnalysis.{AnalysisSourceInfo, AssumptionType, DependencyType}
 import viper.silver.parser.PNode
 import viper.silver.verifier.errors.ErrorNode
 import viper.silver.verifier.{AbstractVerificationError, ConsistencyError, ErrorReason}
@@ -426,100 +421,6 @@ case object Synthesized extends Info {
 abstract class FailureExpectedInfo extends Info {
   override val comment = Nil
   override val isCached = false
-}
-
-object DependencyTypeInfo {
-  def getDependencyTypeInfo(stmt: ast.Stmt): DependencyTypeInfo = {
-    val depType = stmt match {
-      case _: ast.MethodCall => MethodCall
-      case  _: ast.NewStmt | _: ast.AbstractAssign => SourceCode
-      case _: ast.Exhale | _: ast.Assert => ExplicitAssertion
-      case _: ast.Inhale | _: ast.Assume => ExplicitAssumption
-      case _: ast.Fold | _: ast.Unfold | _: ast.Package | _: ast.Apply => Rewrite
-      case _: ast.Quasihavoc | _: ast.Quasihavocall => DependencyType.Implicit
-      case _ => DependencyType.make(AssumptionType.Unknown) /* TODO: should not happen */
-    }
-    DependencyTypeInfo(depType)
-  }
-}
-
-case class DependencyTypeInfo(dependencyType: DependencyType) extends ast.Info {
-
-  override def comment: Seq[String] = Nil
-  override def isCached: Boolean = false
-}
-
-object JoinType extends Enumeration {
-  type JoinType = Value
-  val Source, Sink = Value
-}
-
-object EdgeType extends Enumeration {
-  type EdgeType = Value
-  val Up, Down = Value
-}
-
-/**
- * Represent the piece of information to be stored in dependency nodes in order to join dependency graphs of
- * individual verification components. Nodes with matching join information are connected by an interprocedural edge.
- */
-trait DependencyAnalysisJoinInfo extends ast.Info {
-  override def comment: Seq[String] = Nil
-  override def isCached: Boolean = false
-
-  def matches(dependencyAnalysisJoinInfo: DependencyAnalysisJoinInfo) = {
-    (this, dependencyAnalysisJoinInfo) match {
-      case (SimpleDependencyAnalysisJoin(sourceInfo, joinType, edgeType), SimpleDependencyAnalysisJoin(sourceInfo1, joinType1, edgeType1)) =>
-        sourceInfo.equals(sourceInfo1) && edgeType.equals(edgeType1)
-    }
-  }
-}
-
-case class EvalStackDependencyAnalysisJoin(joinType: JoinType, edgeType: EdgeType) extends DependencyAnalysisJoinInfo
-
-case class SimpleDependencyAnalysisJoin(sourceInfo: AnalysisSourceInfo, joinType: JoinType, edgeType: EdgeType) extends DependencyAnalysisJoinInfo
-
-/**
- * Represent the piece of information to be stored in dependency nodes in order to finalize the intraprocedural
- * dependency graph. Nodes with matching merge information are connected by an intraprocedural edge.
- */
-trait DependencyAnalysisMergeInfo extends ast.Info {
-
-  def isMerge: Boolean
-
-  override def comment: Seq[String] = Nil
-  override def isCached: Boolean = false
-}
-
-case class NoDependencyAnalysisMerge() extends DependencyAnalysisMergeInfo {
-  override def isMerge: Boolean = false
-}
-
-case class SimpleDependencyAnalysisMerge(sourceInfo: AnalysisSourceInfo) extends DependencyAnalysisMergeInfo {
-  override def isMerge: Boolean = true
-}
-
-case class CompositeDependencyAnalysisMergeInfo(sourceInfo1: AnalysisSourceInfo, sourceInfo2: AnalysisSourceInfo) extends DependencyAnalysisMergeInfo {
-	override def isMerge: Boolean = true
-}
-
-object DependencyAnalysisMergeInfo {
-	def attachExpMergeInfo(exps: Seq[ast.Exp], sourceInfo2: Option[AnalysisSourceInfo]): Seq[ast.Exp] = {
-		exps.map(attachExpMergeInfo(_, sourceInfo2))
-	}
-
-	def attachExpMergeInfo(exp: ast.Exp, sourceInfo2: Option[AnalysisSourceInfo]): ast.Exp = {
-		val expSourceInfo = AnalysisSourceInfo.createAnalysisSourceInfo(exp)
-		attachExpMergeInfo(exp, expSourceInfo, sourceInfo2)
-	}
-
-	def attachExpMergeInfo(exp: ast.Exp, sourceInfo1: AnalysisSourceInfo, sourceInfo2: Option[AnalysisSourceInfo]): ast.Exp = {
-		val mergeInfo = if(sourceInfo2.isDefined)
-			CompositeDependencyAnalysisMergeInfo(sourceInfo1, sourceInfo2.get)
-		else
-			SimpleDependencyAnalysisMerge(sourceInfo1)
-		exp.withMeta((exp.pos, ast.MakeInfoPair(mergeInfo, exp.info), exp.errT))
-	}
 }
 
 /** An `Info` instance for composing multiple `Info`s together */
